Do not fix K and MappedActorRef to make the interface more flexible
Let's recap the first solution I architected.

The idea is to allow an actor with a generic interface (that is, ActorRef<M> type) to register to the remoting service by passing a key identifier K (bounded by Hash) and the ActorRef itself.
The user-defined async closure of the remoting service will receive bytes from the network, parse them and compose a Letter<K, MessageTypes> with recipient K and message MessageTypes. Then, the remoting service will get the ActorRef identified by K from a HashMap<K, ActorRef<?>> and dispatch the message to it. But what should the type of this ActorRef be? This is complicated due to the fact that any actor (hence, any ActorRef<M1>, ActorRef<M2>, and so on) should be allowed to register to the service.

Since we cannot know all such types in advance, we need to resort to a "general enough" type.
First, I devised a MappedActorRef<MessageTypes> and made the remoting service accept MappedActorRef<MessageTypes>.
An actor which is willing to register should first convert ("widen") its ActorRef<M> to a general, wider MappedActorRef<MessageTypes> and establish a conversion from MessageTypes to M. The remoting service will therefore send the message through this general interface and not incur in the aforementioned issue about types, which is going to be taken care of by the conversion code.
This is achieved with a trait bound and dynamic dispatch:

```rust
pub fn widen<Supertype>(mut self) -> MappedActorRef<Supertype>
where
	M: TryFrom<Supertype> + Send + 'static,
	Supertype: From<M>,
{
	MappedActorRef {
	    f: Arc::new(Mutex::new(move |supertype: Supertype| {
			let m = supertype.try_into().unwrap_or_else(|_| {
				panic!(
				"conversion failed from {} to {}",
				std::any::type_name::<Supertype>(),
				std::any::type_name::<M>()
				)
			});
			self.try_send(m).map_err(|m| m.into())
	    })),
	}
}
```

There are two downsides:
1. The major downside is that the conversion code in the closure will mostly consist of unwrapping the correct Supertype variant to an M which is useful for the actor identified by K.
2. The implementation of MappedActorRef uses dynamic dispatch.

Can we architect another solution which does not incur in neither downside?

The problem originates from the fact that we used a HashMap<K, ActorRef<MessageTypes>> in the remoting service and accepted ActorRef<MessageTypes>.
What if:
1. The type of the registrant was not fixed and did not require K (that is, was a generic RegistrantType instead), and
2. The type of data structure was not fixed as a HashMap (that is, was a generic RegistrarType instead), and
3. The user could define its own function (RegistrarType, LetterType) for dispatching a letter (LetterType) to a registered actor (a RegistrantType in a RegistrarType)?

RegistrarType should be bounded by a Registrar trait with two methods: register(RegistrantType) and unregister(RegistrantType) (which, in case of a HashMap, will correspond to to insert and remove operations).

We could, for example, group all different ActorRef types in an enum:

```rust
enum MyRegistrantType {
	String(ActorRef<String>),
	U32(ActorRef<u32>),
}
```

First, this would solve the dynamic dispatch problem.

Second, by doing so, we are also removing a lot of rigidity from the architecture:
- Not all subscribers need be identified by a key identifier!
- Sometimes, all subscribers are known in advance, and could (or should) be preloaded and passed in the RegistrarType itself, so that incoming network messages are not lost.
- There could even be just one subscriber, and the HashMap solution would be inefficient.